face_app copy/requirements.txt
flask
flask-cors
pyopenssl          # for self-signed HTTPS cert (camera access)
face-recognition
opencv-python-headless
numpy
Pillow
gunicorn